+++ formats/ladspa/tests/LADSPATestSuite.cpp
@@ -33,12 +33,12 @@ public:
             auto handle = desc->instantiate(desc, sampleRate);
             desc->activate(handle);
             auto portData = std::vector<LADSPA_Data*>(desc->PortCount);
-            for (size_t i = 0; i < desc->PortCount; i++) {
+            for (unsigned long i = 0; i < desc->PortCount; i++) {
                 portData[i] = new LADSPA_Data[sampleRate];
                 desc->connect_port(handle, i, portData[i]);
             }
             desc->run(handle, sampleRate);
-            for (int i = 0; i < desc->PortCount; i++) {
+            for (unsigned long i = 0; i < desc->PortCount; i++) {
                 delete[] portData[i];
             }
             desc->deactivate(handle);
